Remove old class
