+++ app/src/main/AndroidManifest.xml
@@ -22,12 +22,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/WykopAppTheme"
         >
+
         <activity
             android:name="io.github.wykopmobilny.ui.modules.loginscreen.LoginScreenActivity"
-            android:configChanges="keyboardHidden|orientation|screenSize"
-            />
-        <activity
-            android:name="io.github.wykopmobilny.ui.modules.loginscreen.LoginScreenActivityV2"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:theme="@style/Theme.App"
             />
