+++ app/src/main/kotlin/io/github/wykopmobilny/utils/usermanager/UserManager.kt
@@ -5,19 +5,16 @@ import io.github.wykopmobilny.api.responses.LoginResponse
 import io.github.wykopmobilny.storage.api.LoggedUserInfo
 import io.github.wykopmobilny.storage.api.SessionStorage
 import io.github.wykopmobilny.storage.api.UserInfoStorage
-import io.github.wykopmobilny.storage.api.UserSession
 import io.github.wykopmobilny.ui.dialogs.userNotLoggedInDialog
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import javax.inject.Inject
 
-data class LoginCredentials(val login: String, val token: String)
-
 data class UserCredentials(
     val login: String,
     val avatarUrl: String,
     val backgroundUrl: String?,
-    val userKey: String
+    val userKey: String,
 )
 
 interface SimpleUserManagerApi {
@@ -25,7 +22,6 @@ interface SimpleUserManagerApi {
 }
 
 interface UserManagerApi : SimpleUserManagerApi {
-    fun loginUser(credentials: LoginCredentials)
     fun logoutUser()
     fun saveCredentials(credentials: LoginResponse)
     fun runIfLoggedIn(context: Context, callback: () -> Unit)
@@ -37,14 +33,6 @@ class UserManager @Inject constructor(
     private val sessionStorage: SessionStorage,
     private val userInfoStorage: UserInfoStorage,
 ) : UserManagerApi {
-    override fun loginUser(credentials: LoginCredentials) = runBlocking {
-        sessionStorage.updateSession(
-            value = UserSession(
-                login = credentials.login,
-                token = credentials.token,
-            )
-        )
-    }
 
     override fun logoutUser() = runBlocking {
         sessionStorage.updateSession(null)
@@ -58,7 +46,7 @@ class UserManager @Inject constructor(
                 userToken = credentials.userkey,
                 avatarUrl = credentials.profile.avatar,
                 backgroundUrl = credentials.profile.background,
-            )
+            ),
         )
     }
